Handy.List

Description

Handy Functions for manipulating lists.

Documentation

increasingBy :: (a -> a -> Bool) -> [a] -> [a]

Greedily selects ordered subsequence. Usefull for printing progress of optimization search, for example

print . maximum $ possibleSolutions

can be replaced with more entertaining

mapM_ print . increasingBy (<) $ possibleSolutions

ordElem :: Ord a => a -> [a] -> Bool

Test whether element is in ordered list. (Works with infinite lists.)

orderedPermutations :: [a] -> [[a]]

Produces list of permutations. Preserves monotonousBy lt s ==> monotonousBy lt (orderedPermutations s).

This function is slower than permutations, however if you need permurations to be sorted this is more suitable choice as sort (permutations s) not only runs longer than orderedPermutations (sort s) but the latter one is lazy and more memorry efficient.

sort (permutations s) = orderedPermutations (sort s)

orderedSubsets :: [a] -> [[a]]

Produces list-encoded subsets. Preserves monotonousBy lt s ==> monotonousBy lt (orderedSubsets s).

isMonotonousBy :: (a -> a -> Bool) -> [a] -> Bool

Checks monotonicity of list elements by given relation.

isMonotonous :: Ord a => [a] -> Bool

Check monotonicity by default Ord instance.

breakAll :: (a -> Bool) -> [a] -> [[a]]

Break list by delimiters identified by predicate. Variation to words/lines. Works on infinite lists.

>>> head . head $ breakAll (const False) [1..]
1