Handy.Number

Description

Handy functions for manipulating numbers.

Modular arithmetic

egcd :: Integral t => t -> t -> (t, t, t)

Extended Euclidean algorithm (Greatest Common Divisor).

a*x + b*y == g && g == gcd a b where (x,y,g) = egcd a b

modInv :: Integral a => a -> a -> Maybe a

Modular inverse.

(a * i) `mod` m == 1 where Just i = a `modInv` m

Arithmetic

sqrtBig :: Integral b => b -> b

For given n finds largest m such that m^2 <= n. Newtons method.

>>> sqrtBig 121
11
>>> sqrtBig 120
10

Often used seqences

diagonal :: Integral a => [(a, a)]

Diagonal enumeration of pairs of natural numbers. Every pair is guaranteed to occur at finite position in the list. Usable mostly with types representing unbound numbers.

fibs :: [Integer]

List of Fibonacci numbers.

primes :: [Integer]

List of prime numbers.

Factorization

factor :: Integer -> [Integer]

Reasonably fast, but due to the fact that it uses rho function for numbers not divisible by first 1000 primes, resulting list may contain also composites :-(.

Encoding/Decoding

enc1 :: Integral a => a -> a -> a

Encode tuple of natural numbers into position on diagonal enumeration.

map (uncurry enc1) diagonal == [0..]

dec1 :: Integral a => a -> (a, a)

Decoding for enc1 encoding.

For natural numbers:

dec1 . uncurry enc1 = id

uncurry enc1 . dec1 = id

enc2 :: Integral a => a -> a -> a

Alternate encoding of tuples of natural numbers.

dec2 :: Integral a => a -> (a, a)

Decoding for enc2 encoding. For natural numbers:

dec2 . uncurry enc2 = id

uncurry enc2 . dec2 = id

encodeBase :: Integral a => [b] -> a -> [b]

Encode natural number in base (length s), using symbols from s. Provided s must be of length no less than 2.

>>> encodeBase ['0'..'9'] 100
"100"
>>> encodeBase "01" 5
"101"
>>> encodeBase "0123456789abc" (6*9)
"42"

It is not a bijection due to "leading zeros problem".

decodeBase :: (Integral a, Eq b) => [b] -> [b] -> a

Decode natural number in base (length s), using symbols from s. Provided s must be of length no less than 2. It is dual to encodeBase.

>>> decodeBase ['0'..'9'] "101" :: Int
101
>>> encodeBase "01" "101" :: Int
5

It is not a bijection due to "leading zeros problem".

decodeBase s . encodeBase s = id

encodeBij :: Integral a => [b] -> a -> [b]

Bijectively encode natural numbers into sequence. This is not standard math encoding, but is handy when you need to use space as efficiently as you can.

Provided s must be non-empty. For single element list it collapses to length-encoding.

Handy for generating examples:

map (encodeBij "aA") [1..]

decodeBij :: (Integral a, Eq b) => [b] -> [b] -> a

Decoding for encodeBij encoding.

For natural numbers and non-empty list:

encodeBij s . decodeBij s = id

decodeBij s . encodeBij s = id

Convenience functions/values

binBase :: [Char]

List of symbols for binary encoding. For use with encodeBase and decodeBase.

octBase :: [Char]

List of symbols for octal encoding. For use with encodeBase and decodeBase.

decBase :: [Char]

List of symbols for decimal encoding. For use with encodeBase and decodeBase.

hexBase :: [Char]

List of symbols for hexadecimal encoding. For use with encodeBase and decodeBase.